Add qiskit.opflow and qiskit.algorithms to docs (Qiskit/qiskit#5477)
* Add qiskit.opflow to docs

In Qiskit/qiskit#5460 we added a new module qiskit.opflow which was a migration of
the operator flow from qiskit-aqua into qiskit-terra. That PR was just a
straight commit history migration and didn't include the new module in
the documentation. This commit fixes this so that we include the opflow
documentation in the builds.

* Try to fix docs build

* Try to remove docstring from abstract method

* Remove permute from statefn

* Add algorithms to docs too

* Revert "Remove permute from statefn"

This reverts commit 2b559ce70d4736db5f161b50ab907c19f64e9627.

* Revert "Try to remove docstring from abstract method"

This reverts commit 09cf825369ce742f5f4a75b2d84c302eb53c2ffb.

* Fix algorithm error docs typo

* Fix state_fn abstract method failures

* Add more rst files for specific subclass fields

* Add missing ref tags

* Fix build warning on duplicate Instruction

* Revert changes to abstract state_fn docs

Co-authored-by: Julien Gacon <gaconju@gmail.com>