Upgrade to pylint 2.7 (Qiskit/qiskit#5630)
* upgrade astroid, pylint

* update for pylint 2.6

* Apply suggestions from code review

* update to pylint 2.6.2

* upgrade pylint 2.6.2 in requirements-dev.txt

* Bump to latest pylint version 2.7.0

This commit bumps the pylint version to pylint 2.7.0 which adds python
3.9 compatibility. It fixes issues identified by new and changed rules
in this new version. These fall into 3 primary categories,
'use-a-generator' fixes which remove intermediate lists or other
containers inside a generator function, disables in places where pylint
incorrectly thinks a QuantumCircuit is list and fails on no-member, and
disables in places where pylint is unable to correctly resolve a kwarg.

* Fix new errors from rebase

* Upgrade to 2.7.1

* revert unnecessary change

* Revert

* fix linter error

* fix linter error

Co-authored-by: Matthew Treinish <mtreinish@kortar.org>
Co-authored-by: Manoel Marques <Manoel.Marques@ibm.com>