Stop using networkx for qaoa tests (Qiskit/qiskit#6142)
* Stop using networkx for qaoa tests

In the qaoa tests networkx was being used to generate a random
adjancency matrix as an input. This would work fine especially as the
graph size is small so performance isn't an issue. However, with the
recent release of decorator 5.0.0 earlier today these tests are
failing because of an internal incompatibility with the random state
decorator networkx is using for random_gnp_graph() function. Instead
of pinning decorator in the constraints file and hoping people running
tests without tox know to use it, this commit just switches that test to
use retworkx instead. retworkx doesn't have an external dependency on
decorator (or anything except numpy) and can generate an adjacency
matrix for a gnp random graph in basically the same exact way without
this incompatibility.

* Add decorator to constraints too

The same networkx example is used in the tutorial notebooks, since we
don't control the source to those (and we probably want to continue to
use networkx there since it's a popular package) this adds decorator to
the constraints file too to unblock the tutorials job.