[WIP] Change attribute `most_likely_phase` to `phase` in PhaseEstimator (Qiskit/qiskit#6238)
* Change attribute `most_likely_phase` to `phase` in PhaseEstimator

IterativePhaseEstimation Qiskit/qiskit#6096 will satisfy the PhaseEstimator interface.
However, the property for retrieving the phase is currently
`PhaseEstimator.most_likely_phase`, which is too specific to accomodate
IterativePhaseEstimation, which only computes a single phase in any
case. This PR renames this property to `phase`, and leaves the
definition of `phase` to the specific implementation.

* Include deprecated methods most_likely_phase

* Add version number to deprecation messages

* Fix for linter complaint