Fix Qiskit/qiskit#7142 (Qiskit/qiskit#7144)
* Fix Qiskit/qiskit#7142

Ensures a unified API of the `aux_operator_eigenvalues`.
Prior to this commit, the returned types of the `NumPyEigensolver` and
`VQE` were mismatched. Furthermore, the documented type hint in the
`EigensolverResult` was yet another value.
This commit unifies the API, documents the expected type properly and
adds unittests which assert that this bug remains fixed.

* Do not map expectation value to real

* Compute aux operator standard deviations during VQE

* Assert standard deviations in EigensolverResult objects

* Improve aux_operator_eigenvalues docstring

* Fix NumPyEigensolver threshold check

* Run black

* Update release note

* Avoid ambiguous truth value of numpy arrays

* Add aux_op test for qasm-based VQE (non-trivial std dev)

* Remove erroneous nesting

This was accidentally copied from the `EigensolverResult` case.

* Remove release note

This fixes a bug in an unreleased feature so this release note is
actually not needed (in fact its more confusing than helpful).

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>