Skip classical function tests if tweedledum isn't present (Qiskit/qiskit#8876)
* Skip classical function tests if tweedledum isn't present

This commit fixes an oversight in Qiskit/qiskit#8849 and Qiskit/qiskit#8818 which made tweedledum
optional for arm64 macOS systems which was the unittest suite still
unconditionally depends on tweedledum being installed to run the
classical function compiler tests. This commit fixes this by making the
tests skip if tweedledum isn't installed so that M1 mac users are able
to run the full test suite again.

* Skip other tests requiring tweedledum