Switch from retworkx to rustworkx package (Qiskit/qiskit#9162)
* Switch from retworkx to rustworkx package

In the retworkx/rustworkx 0.12.0 release the package was renamed from
retworkx to rustworkx which was done at the request of the networkx
maintainers [1]. While the retworkx name continues to work for backwards
compatibility, it is best to move off of the legacy name and start using
the new name sooner rather than later. This commit updates all the
retworkx usage to the new rustworkx and updates our requirements list to
directly depend on rustworkx.

[1] https://qiskit.org/documentation/rustworkx/release_notes.html#prelude

* Fix release note path

* Update releasenotes/notes/rustworkx-not-retworkx-b7c4da600df58701.yaml

Co-authored-by: Jake Lishman <jake@binhbar.com>

Co-authored-by: Jake Lishman <jake@binhbar.com>
Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>